Break out persisted.rs
