+++ crates/bin/view/src/main.rs
@@ -1,6 +1,8 @@
 mod opt;
+mod persisted;
 
 use opt::*;
+use persisted::*;
 
 use anyhow::Context;
 
@@ -30,152 +32,6 @@ struct SceneInstanceDesc {
     #[serde(default = "default_instance_scale")]
     scale: [f32; 3],
     mesh: String,
-}
-
-#[derive(Clone, PartialEq, serde::Serialize, serde::Deserialize)]
-struct SunState {
-    theta: f32,
-    phi: f32,
-}
-
-impl SunState {
-    pub fn direction(&self) -> Vec3 {
-        fn spherical_to_cartesian(theta: f32, phi: f32) -> Vec3 {
-            let x = phi.sin() * theta.cos();
-            let y = phi.cos();
-            let z = phi.sin() * theta.sin();
-            Vec3::new(x, y, z)
-        }
-
-        spherical_to_cartesian(self.theta, self.phi)
-    }
-}
-
-#[derive(Clone, PartialEq, serde::Serialize, serde::Deserialize)]
-struct LocalLightsState {
-    theta: f32,
-    phi: f32,
-    count: u32,
-    distance: f32,
-    multiplier: f32,
-}
-
-fn always_true() -> bool {
-    true
-}
-
-fn default_camera_speed() -> f32 {
-    2.5
-}
-
-fn default_camera_smoothness() -> f32 {
-    1.0
-}
-
-fn default_sun_rotation_smoothness() -> f32 {
-    1.0
-}
-
-trait ShouldResetPathTracer {
-    fn should_reset_path_tracer(&self, other: &Self) -> bool;
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-struct CameraState {
-    position: Vec3,
-    rotation: Quat,
-    vertical_fov: f32,
-}
-
-impl Default for CameraState {
-    fn default() -> Self {
-        Self {
-            position: Vec3::ONE,
-            rotation: Quat::IDENTITY,
-            vertical_fov: 62.0,
-        }
-    }
-}
-
-impl ShouldResetPathTracer for CameraState {
-    fn should_reset_path_tracer(&self, other: &Self) -> bool {
-        !self.position.abs_diff_eq(other.position, 1e-5)
-            || !self.rotation.abs_diff_eq(other.rotation, 1e-5)
-            || self.vertical_fov != other.vertical_fov
-    }
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-struct LightState {
-    emissive_multiplier: f32,
-    #[serde(default = "always_true")]
-    enable_emissive: bool,
-    sun: SunState,
-    local_lights: LocalLightsState,
-}
-
-impl Default for LightState {
-    fn default() -> Self {
-        Self {
-            emissive_multiplier: 1.0,
-            enable_emissive: true,
-            sun: SunState {
-                theta: -4.54,
-                phi: 1.48,
-            },
-            local_lights: LocalLightsState {
-                theta: 1.0,
-                phi: 1.0,
-                count: 0,
-                distance: 1.5,
-                multiplier: 10.0,
-            },
-        }
-    }
-}
-
-impl ShouldResetPathTracer for LightState {
-    fn should_reset_path_tracer(&self, other: &Self) -> bool {
-        self.emissive_multiplier != other.emissive_multiplier
-            || self.sun != other.sun
-            || self.local_lights != other.local_lights
-    }
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-struct MovementState {
-    #[serde(default = "default_camera_speed")]
-    camera_speed: f32,
-    #[serde(default = "default_camera_smoothness")]
-    camera_smoothness: f32,
-    #[serde(default = "default_sun_rotation_smoothness")]
-    sun_rotation_smoothness: f32,
-}
-
-impl Default for MovementState {
-    fn default() -> Self {
-        Self {
-            camera_speed: 1.0,
-            camera_smoothness: default_camera_smoothness(),
-            sun_rotation_smoothness: default_sun_rotation_smoothness(),
-        }
-    }
-}
-
-#[derive(Clone, serde::Serialize, serde::Deserialize)]
-struct ExposureState {
-    ev_shift: f32,
-    #[serde(default)]
-    use_dynamic_adaptation: bool,
-}
-
-impl Default for ExposureState {
-    fn default() -> Self {
-        Self {
-            ev_shift: 0.0,
-            use_dynamic_adaptation: true,
-        }
-    }
 }
 
 struct RuntimeState {
